+++ src/main/java/com/raoulvdberge/refinedstorage/gui/sidebutton/SideButtonGridViewType.java
@@ -2,7 +2,6 @@ package com.raoulvdberge.refinedstorage.gui.sidebutton;
 
 import com.raoulvdberge.refinedstorage.api.network.grid.IGrid;
 import com.raoulvdberge.refinedstorage.gui.GuiBase;
-import com.raoulvdberge.refinedstorage.tile.grid.portable.IPortableGrid;
 import net.minecraft.util.text.TextFormatting;
 
 public class SideButtonGridViewType extends SideButton {
@@ -28,26 +27,14 @@ public class SideButtonGridViewType extends SideButton {
     public void actionPerformed() {
         int type = grid.getViewType();
 
-        if (grid instanceof IPortableGrid) {
-            if (type == IGrid.VIEW_TYPE_NORMAL_DETAILED) {
-                type = IGrid.VIEW_TYPE_NORMAL;
-            } else {
-                type = IGrid.VIEW_TYPE_NORMAL_DETAILED;
-            }
+        if (type == IGrid.VIEW_TYPE_NORMAL) {
+            type = IGrid.VIEW_TYPE_NON_CRAFTABLES;
+        } else if (type == IGrid.VIEW_TYPE_NON_CRAFTABLES) {
+            type = IGrid.VIEW_TYPE_CRAFTABLES;
+        } else if (type == IGrid.VIEW_TYPE_CRAFTABLES) {
+            type = IGrid.VIEW_TYPE_NORMAL;
         } else {
-            if (type == IGrid.VIEW_TYPE_NORMAL_DETAILED) {
-                type = IGrid.VIEW_TYPE_NON_CRAFTABLES_DETAILED;
-            } else if (type == IGrid.VIEW_TYPE_NON_CRAFTABLES_DETAILED) {
-                type = IGrid.VIEW_TYPE_CRAFTABLES_DETAILED;
-            } else if (type == IGrid.VIEW_TYPE_CRAFTABLES_DETAILED) {
-                type = IGrid.VIEW_TYPE_NORMAL;
-            } else if (type == IGrid.VIEW_TYPE_NORMAL) {
-                type = IGrid.VIEW_TYPE_NON_CRAFTABLES;
-            } else if (type == IGrid.VIEW_TYPE_NON_CRAFTABLES) {
-                type = IGrid.VIEW_TYPE_CRAFTABLES;
-            } else if (type == IGrid.VIEW_TYPE_CRAFTABLES) {
-                type = IGrid.VIEW_TYPE_NORMAL_DETAILED;
-            }
+            type = IGrid.VIEW_TYPE_NORMAL; // @todo: Remove in 1.13 (1.5.24 -> 1.5.26 conversion, removed detailed grid view type)
         }
 
         grid.onViewTypeChanged(type);